# Speed up numerai_corr x 2