This PR introduces a benchmark for the power function and numerai_corr, running each function 1000 times to measure performance. Additionally, a significant speedup was achieved by optimizing a slow assertion in the power function, replacing the power function it with a more efficient numpy based implementation.

The results of the speedup are: 

Master:
Execution time power_bench: 3.6272 seconds
Execution time numerai_corr_bench: 10.5109 seconds

This branch:
Execution time power_bench: 0.7179 seconds
Execution time numerai_corr_bench: 4.9527 seconds



To run the benchmark do: 
python -m  tests.test_scoring
Additionally, it can be run with --profile argument to do profiling of the functions.